requirements.txt
asgiref==3.7.2
Django==3.2.19
djangorestframework==3.15.1
pytz==2023.3
sqlparse==0.4.4
